adicionado método preliminar de contagem de grãos por classe

+++ Funcoes/gerarHistograma.m
@@ -1,6 +1,6 @@
 function gerarHistograma( imagem )
 
-    hist(imagem);
+    hist(double(imagem));
      
 end
